requirements.txt
slackclient==1.3.1
slackeventsapi==2.2.1
Flask>=1.1.1